-Added verification for darwin aliases.

+++ sys/shared/.config/nushell/config.nu
@@ -608,12 +608,15 @@ def kr [input: string] {
 alias ks = kotlinc-jvm
 
 # Xcode
-alias kx = killall Xcode
-alias repd = xcodebuild -resolvePackageDependencies
-alias rmdd = rm -rf ~/Library/Developer/Xcode/DerivedData
-alias rmspm = rm -rf ~/Library/org.swift.swiftpm/
-alias rmcspm = rm -rf ~/Library/Caches/org.swift.swiftpm/
-alias xo = xed .
+let osName = (sys | get host | get name)
+if $osName == 'Darwin' {
+	alias kx = killall Xcode
+	alias repd = xcodebuild -resolvePackageDependencies
+	alias rmdd = rm -rf ~/Library/Developer/Xcode/DerivedData
+	alias rmspm = rm -rf ~/Library/org.swift.swiftpm/
+	alias rmcspm = rm -rf ~/Library/Caches/org.swift.swiftpm/
+	alias xo = xed .
+}
 
 source ~/.cache/starship/init.nu
